refactor(core): 使用Long替换BigInteger优化ID类型
- 将所有实体和相关接口中的ID类型从BigInteger更改为Long
- 修改序列化相关配置，确保Long类型ID正确处理
- 更新测试代码中的ID类型为Long，修正上下文写入部分
- 重构部分工具类，实现Long与UUID的转换支持
- 增加新的application配置，扩展dashscope及相关模型配置
- 修正代码注释和方法签名，提升代码可读性和一致性
- 删除不再使用的BigInteger导入及相关代码引用
- 新增WebClient测试方法验证recommend接口连通性

+++ common-utils-cloud/db-table-spring-boot-autoconfigure/src/main/java/com/db/cursorQuery/MongoCursorQuery.java
@@ -12,7 +12,7 @@ import org.springframework.data.mongodb.core.query.Query;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.math.BigInteger;
+
 import java.util.Collection;
 import java.util.List;
 
@@ -139,7 +139,7 @@ public class MongoCursorQuery<T> {
     }
 
     public Flux<T> list() {
-        BigInteger lastId = validatedPage.getLastId();
+        Long lastId = validatedPage.getLastId();
         int pageSize = validatedPage.getPageSize();
 
         Criteria cursorCriteria = criteria;
