+++ common-utils-cloud/public-pojo-spring-boot-autoconfigure/src/main/java/com/guanshiyun/profile/SKUApiVO.java
@@ -14,7 +14,7 @@ import lombok.experimental.SuperBuilder;
 import java.io.Serial;
 import java.io.Serializable;
 import java.math.BigDecimal;
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -28,7 +28,7 @@ public class SKUApiVO extends BasePojo implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
     @JsonSerialize(using = ToStringSerializer.class)
-    private BigInteger id;
+    private Long id;
     //商品id
 
     /** 名称*/
@@ -67,7 +67,7 @@ public class SKUApiVO extends BasePojo implements Serializable {
     // 商品详情
     private String detailContent;
     //商品id
-    private BigInteger productId;
+    private Long productId;
     //图片
     private List<String> picList;
 }
