+++ common-utils-cloud/public-pojo-spring-boot-autoconfigure/src/main/java/com/guanshiyun/responsepojo/CursorPageResult.java
@@ -3,7 +3,7 @@ package com.guanshiyun.responsepojo;
 import lombok.*;
 import lombok.experimental.Accessors;
 
-import java.math.BigInteger;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -11,13 +11,13 @@ import java.math.BigInteger;
 @Accessors(chain = true)
 public class CursorPageResult<T> {
     //lastId
-    private BigInteger cursor;
+    private Long cursor;
     //结果集
     private  T rows;
     //是否存在下一页
     private Boolean hasNext;
 
-    public static <T> CursorPageResult<T> of(BigInteger cursor, T rows, Boolean hasNext) {
+    public static <T> CursorPageResult<T> of(Long cursor, T rows, Boolean hasNext) {
         return CursorPageResult.<T>builder()
                 .cursor(cursor)
                 .rows(rows)
