+++ gsy-ai-cloud/ai-pojo/src/main/java/com/guanshiyun/knowledge/KnowledgeRepository.java
@@ -5,7 +5,7 @@ import lombok.experimental.FieldNameConstants;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
 
-import java.math.BigInteger;
+
 import java.time.LocalDateTime;
 
 /**
@@ -19,16 +19,16 @@ import java.time.LocalDateTime;
 @Table("knowledge_repository")
 public class KnowledgeRepository {
     @Id
-    private BigInteger id;
+    private Long id;
     /**
      * 创建者，目前使用 SysUser 的 id 编号
      *
      */
-    public BigInteger creator;
+    public Long creator;
     /**
      * 更新者，目前使用 SysUser 的 id 编号
      */
-    public BigInteger updater;
+    public Long updater;
     /**
      * 创建时间
      */
