+++ gsy-ai-cloud/ai-server/src/main/java/com/guanshiyun/repository/chat/ChatRecordRepository.java
@@ -3,7 +3,7 @@ package com.guanshiyun.repository.chat;
 import com.guanshiyun.chat.ChatRecord;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-import java.math.BigInteger;
 
-public interface ChatRecordRepository extends ReactiveCrudRepository<ChatRecord, BigInteger> {
+
+public interface ChatRecordRepository extends ReactiveCrudRepository<ChatRecord, Long> {
 }
