+++ gsy-ai-cloud/ai-server/src/main/java/com/guanshiyun/repository/embedding/ActiveSimilarityThresholdConfigurationRepository.java
@@ -3,7 +3,7 @@ package com.guanshiyun.repository.embedding;
 import com.guanshiyun.embedding.ActiveSimilarityThresholdConfiguration;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
-import java.math.BigInteger;
 
-public interface ActiveSimilarityThresholdConfigurationRepository extends R2dbcRepository<ActiveSimilarityThresholdConfiguration, BigInteger> {
+
+public interface ActiveSimilarityThresholdConfigurationRepository extends R2dbcRepository<ActiveSimilarityThresholdConfiguration, Long> {
 }
